Moved runtime library to global dictionary
